+++ thrift/src/test/scala/com/twitter/finatra/thrift/tests/LegacyDoEverythingThriftServerFeatureTest.scala
@@ -2,7 +2,6 @@ package com.twitter.finatra.thrift.tests
 
 import com.twitter.conversions.DurationOps._
 import com.twitter.doeverything.thriftscala.{Answer, DoEverything, Question}
-import com.twitter.finagle.http.Status
 import com.twitter.finagle.tracing.Trace
 import com.twitter.finagle.{Filter, Service}
 import com.twitter.finatra.thrift.EmbeddedThriftServer
@@ -13,9 +12,10 @@ import com.twitter.io.Buf
 import com.twitter.scrooge
 import com.twitter.util.{Await, Future}
 import org.apache.thrift.TApplicationException
-import scala.util.parsing.json.JSON
 
-@deprecated("These tests exist to ensure legacy functionaly still operates. Do not use them for guidance", "2018-12-20")
+@deprecated(
+  "These tests exist to ensure legacy functionality still operates. Do not use them for guidance",
+  "2018-12-20")
 class LegacyDoEverythingThriftServerFeatureTest extends FeatureTest {
   override val server = new EmbeddedThriftServer(
     twitterServer = new LegacyDoEverythingThriftServer,
@@ -24,7 +24,8 @@ class LegacyDoEverythingThriftServerFeatureTest extends FeatureTest {
   )
 
   /* Higher-kinded interface type */
-  val client123: DoEverything[Future] = server.thriftClient[DoEverything[Future]](clientId = "client123")
+  val client123: DoEverything[Future] =
+    server.thriftClient[DoEverything[Future]](clientId = "client123")
   /* Method-Per-Endpoint type: https://twitter.github.io/scrooge/Finagle.html#id1 */
   val methodPerEndpointClient123: DoEverything.MethodPerEndpoint =
     server.thriftClient[DoEverything.MethodPerEndpoint](clientId = "client123")
@@ -84,16 +85,22 @@ class LegacyDoEverythingThriftServerFeatureTest extends FeatureTest {
     val service = filter.andThen(servicePerEndpoint123.uppercase)
     await(service(DoEverything.Uppercase.Args("hello"))) should equal("GOODBYE")
 
-    val filter2 = new Filter[scrooge.Request[DoEverything.Uppercase.Args], scrooge.Response[
-      DoEverything.Uppercase.SuccessType
-    ], scrooge.Request[DoEverything.Uppercase.Args], scrooge.Response[
-      DoEverything.Uppercase.SuccessType
-    ]] {
+    val filter2 = new Filter[
+      scrooge.Request[DoEverything.Uppercase.Args],
+      scrooge.Response[
+        DoEverything.Uppercase.SuccessType
+      ],
+      scrooge.Request[DoEverything.Uppercase.Args],
+      scrooge.Response[
+        DoEverything.Uppercase.SuccessType
+      ]] {
       override def apply(
         request: scrooge.Request[DoEverything.Uppercase.Args],
-        service: Service[scrooge.Request[DoEverything.Uppercase.Args], scrooge.Response[
-          DoEverything.Uppercase.SuccessType
-        ]]
+        service: Service[
+          scrooge.Request[DoEverything.Uppercase.Args],
+          scrooge.Response[
+            DoEverything.Uppercase.SuccessType
+          ]]
       ): Future[scrooge.Response[DoEverything.Uppercase.SuccessType]] = {
         val filteredRequest: scrooge.Request[DoEverything.Uppercase.Args] =
           scrooge.Request(Map("com.twitter.test.header" -> Seq(Buf.Utf8("foo"))), request.args)
@@ -246,59 +253,28 @@ class LegacyDoEverythingThriftServerFeatureTest extends FeatureTest {
     MDC.get("traceId") should be(traceId.traceId.toString())
   }
 
-  test("GET /admin/registry.json") {
-    val response = server.httpGetAdmin(
-      "/admin/registry.json",
-      andExpect = Status.Ok)
-
-    val json: Map[String, Any] =
-      JSON.parseFull(response.contentString).get.asInstanceOf[Map[String, Any]]
-
-    val registry = json("registry").asInstanceOf[Map[String, Any]]
-    registry.contains("library") should be(true)
-    registry("library").asInstanceOf[Map[String, String]].contains("finatra") should be(true)
-
-    val finatra = registry("library")
-      .asInstanceOf[Map[String, Any]]("finatra")
-      .asInstanceOf[Map[String, Any]]
-
-    finatra.contains("thrift") should be(true)
-    val thrift = finatra("thrift").asInstanceOf[Map[String, Any]]
-    thrift.contains("filters") should be(true)
-    thrift.contains("methods") should be(true)
-
-    val methods = thrift("methods").asInstanceOf[Map[String, Any]]
-    methods.size should be > 0
-
-    methods.foreach { case (_, data) =>
-      data.isInstanceOf[Map[_, _]] should be(true)
-      val methodJsonInformation = data.asInstanceOf[Map[String, Any]]
-      methodJsonInformation.contains("service_name") should be(true)
-      methodJsonInformation.contains("class") should be(true)
-    }
-  }
-
   test("Basic server stats") {
     await(client123.uppercase("Hi")) should equal("HI")
-    server.assertCounter("srv/thrift/sent_bytes")(_ > 0)
-    server.assertCounter("srv/thrift/received_bytes")(_ > 0)
-    server.assertCounter("srv/thrift/requests", 1L)
-    server.assertCounter("srv/thrift/success", 1L)
+    server.inMemoryStats.counters.assert("srv/thrift/sent_bytes")(_ > 0)
+    server.inMemoryStats.counters.assert("srv/thrift/received_bytes")(_ > 0)
+    server.inMemoryStats.counters.assert("srv/thrift/requests", 1)
+    server.inMemoryStats.counters.assert("srv/thrift/success", 1)
   }
 
   test("Per-method stats scope") {
     val question = Question("fail")
     await(client123.ask(question)) should equal(Answer("DoEverythingException caught"))
-    server.assertCounter("per_method_stats/ask/success", 1L)
-    server.assertCounter("per_method_stats/ask/failures", 0L)
+    server.inMemoryStats.counters.assert("per_method_stats/ask/success", 1)
+    server.inMemoryStats.counters.get("per_method_stats/ask/failures") should be(
+      None)
   }
 
   test("Per-endpoint stats scope") {
     val question = Question("fail")
     await(client123.ask(question)) should equal(Answer("DoEverythingException caught"))
-    server.assertCounter("srv/thrift/ask/requests", 1L)
-    server.assertCounter("srv/thrift/ask/success", 1L)
-    server.assertCounter("srv/thrift/ask/failures", 0L)
+    server.inMemoryStats.counters.assert("srv/thrift/ask/requests", 1)
+    server.inMemoryStats.counters.assert("srv/thrift/ask/success", 1)
+    server.inMemoryStats.counters.assert("srv/thrift/ask/failures", 0)
   }
 
   private def await[T](f: Future[T]): T = {